+++ lib/crossbeams/layout/renderer/checkbox.rb
@@ -10,6 +10,7 @@ module Crossbeams
           @field_config = field_config
           @page_config  = page_config
           @caption      = field_config[:caption] || present_field_as_label(field_name)
+          @tooltip      = field_config[:tooltip]
         end
 
         def render # rubocop:disable Metrics/AbcSize
@@ -20,7 +21,7 @@ module Crossbeams
           <div #{wrapper_id} class="#{div_class}">#{hint_text}
             <input #{name_attribute} type="hidden" value="f">
             <input type="checkbox" value="t" #{checked} #{name_attribute} #{field_id} #{attrs.join(' ')}>
-            <label for="#{@page_config.name}_#{@field_name}">#{@caption}#{error_state}</label>
+            <label for="#{@page_config.name}_#{@field_name}"#{tooltip}>#{@caption}#{error_state}</label>
             <div class="order-1">#{hint_trigger}</div>
           </div>
           HTML
@@ -34,6 +35,11 @@ module Crossbeams
 
         def checked
           value && value != false && value != 'f' && value != 'false' && value.to_s != '0' ? 'checked' : ''
+        end
+
+        def tooltip
+          return '' unless @tooltip
+          %( title="#{@tooltip}")
         end
       end
     end